add new middlewares
