+++ middlewares/index.js
@@ -1,7 +1,12 @@
-const getAllGames = require("./games");
+const { getAllGames, checkIsTitleInArray, updateGamesArray, updateGamesFile, findGameById, deleteGame } = require("./games");
 const cors = require("./cors");
 
 module.exports = {
     getAllGames,
+    checkIsTitleInArray,
+    updateGamesArray,
+    updateGamesFile,
+    findGameById,
+    deleteGame,
     cors
 }
