+++ routes/games.js
@@ -1,9 +1,9 @@
 const gamesRouter = require("express").Router();
-const { sendAllGames, deleteGame, addGameController } = require("../controllers/games");
-const { getAllGames } = require("../middlewares");
+const { sendAllGames, sendUpdatedGames } = require("../controllers/games");
+const { getAllGames, checkIsTitleInArray, updateGamesArray, updateGamesFile, findGameById, deleteGame } = require("../middlewares");
 
 gamesRouter.get("/games", getAllGames, sendAllGames);
-gamesRouter.delete("/games/:id", getAllGames, deleteGame);
-gamesRouter.post("/games", getAllGames, addGameController)
+gamesRouter.post("/games", getAllGames, checkIsTitleInArray, updateGamesArray, updateGamesFile, sendUpdatedGames);
+gamesRouter.delete("/games/:id", getAllGames, findGameById, deleteGame, updateGamesFile, sendUpdatedGames);
 
 module.exports = gamesRouter;
