+++ routes/index.js
@@ -1,0 +1,7 @@
+const mainRoute = require("./main");
+const gamesRouter = require("./games");
+
+module.exports = {
+    mainRoute,
+    gamesRouter
+}
